+++ R/detect.R
@@ -42,13 +42,31 @@ str_detect <- function(string, pattern, negate = FALSE) {
   check_lengths(string, pattern)
   check_bool(negate)
 
-  switch(type(pattern),
+  out <- switch(
+    type(pattern),
     empty = no_empty(),
     bound = no_boundary(),
-    fixed = stri_detect_fixed(string, pattern, negate = negate, opts_fixed = opts(pattern)),
-    coll  = stri_detect_coll(string,  pattern, negate = negate, opts_collator = opts(pattern)),
-    regex = stri_detect_regex(string, pattern, negate = negate, opts_regex = opts(pattern))
+    fixed = stri_detect_fixed(
+      string,
+      pattern,
+      negate = negate,
+      opts_fixed = opts(pattern)
+    ),
+    coll = stri_detect_coll(
+      string,
+      pattern,
+      negate = negate,
+      opts_collator = opts(pattern)
+    ),
+    regex = stri_detect_regex(
+      string,
+      pattern,
+      negate = negate,
+      opts_regex = opts(pattern)
+    )
   )
+
+  preserve_names_if_possible(string, pattern, out)
 }
 
 #' Detect the presence/absence of a match at the start/end
@@ -78,16 +96,33 @@ str_starts <- function(string, pattern, negate = FALSE) {
   check_lengths(string, pattern)
   check_bool(negate)
 
-  switch(type(pattern),
+  out <- switch(
+    type(pattern),
     empty = no_empty(),
     bound = no_boundary(),
-    fixed = stri_startswith_fixed(string, pattern, negate = negate, opts_fixed = opts(pattern)),
-    coll  = stri_startswith_coll(string, pattern, negate = negate, opts_collator = opts(pattern)),
+    fixed = stri_startswith_fixed(
+      string,
+      pattern,
+      negate = negate,
+      opts_fixed = opts(pattern)
+    ),
+    coll = stri_startswith_coll(
+      string,
+      pattern,
+      negate = negate,
+      opts_collator = opts(pattern)
+    ),
     regex = {
       pattern2 <- paste0("^(", pattern, ")")
-      stri_detect_regex(string, pattern2, negate = negate, opts_regex = opts(pattern))
+      stri_detect_regex(
+        string,
+        pattern2,
+        negate = negate,
+        opts_regex = opts(pattern)
+      )
     }
   )
+  preserve_names_if_possible(string, pattern, out)
 }
 
 #' @rdname str_starts
@@ -96,16 +131,33 @@ str_ends <- function(string, pattern, negate = FALSE) {
   check_lengths(string, pattern)
   check_bool(negate)
 
-  switch(type(pattern),
+  out <- switch(
+    type(pattern),
     empty = no_empty(),
     bound = no_boundary(),
-    fixed = stri_endswith_fixed(string, pattern, negate = negate, opts_fixed = opts(pattern)),
-    coll  = stri_endswith_coll(string, pattern, negate = negate, opts_collator = opts(pattern)),
+    fixed = stri_endswith_fixed(
+      string,
+      pattern,
+      negate = negate,
+      opts_fixed = opts(pattern)
+    ),
+    coll = stri_endswith_coll(
+      string,
+      pattern,
+      negate = negate,
+      opts_collator = opts(pattern)
+    ),
     regex = {
       pattern2 <- paste0("(", pattern, ")$")
-      stri_detect_regex(string, pattern2, negate = negate, opts_regex = opts(pattern))
+      stri_detect_regex(
+        string,
+        pattern2,
+        negate = negate,
+        opts_regex = opts(pattern)
+      )
     }
   )
+  preserve_names_if_possible(string, pattern, out)
 }
 
 #' Detect a pattern in the same way as `SQL`'s `LIKE` and `ILIKE` operators
@@ -148,7 +200,9 @@ str_like <- function(string, pattern, ignore_case = deprecated()) {
   check_lengths(string, pattern)
   check_character(pattern)
   if (inherits(pattern, "stringr_pattern")) {
-    cli::cli_abort("{.arg pattern} must be a plain string, not a stringr modifier.")
+    cli::cli_abort(
+      "{.arg pattern} must be a plain string, not a stringr modifier."
+    )
   }
   if (lifecycle::is_present(ignore_case)) {
     lifecycle::deprecate_warn(
@@ -166,7 +220,8 @@ str_like <- function(string, pattern, ignore_case = deprecated()) {
   }
 
   pattern <- regex(like_to_regex(pattern), ignore_case = FALSE)
-  stri_detect_regex(string, pattern, opts_regex = opts(pattern))
+  out <- stri_detect_regex(string, pattern, opts_regex = opts(pattern))
+  preserve_names_if_possible(string, pattern, out)
 }
 
 #' @export
@@ -175,15 +230,22 @@ str_ilike <- function(string, pattern) {
   check_lengths(string, pattern)
   check_character(pattern)
   if (inherits(pattern, "stringr_pattern")) {
-    cli::cli_abort(tr_("{.arg pattern} must be a plain string, not a stringr modifier."))
+    cli::cli_abort(tr_(
+      "{.arg pattern} must be a plain string, not a stringr modifier."
+    ))
   }
 
   pattern <- regex(like_to_regex(pattern), ignore_case = TRUE)
-  stri_detect_regex(string, pattern, opts_regex = opts(pattern))
+  out <- stri_detect_regex(string, pattern, opts_regex = opts(pattern))
+  preserve_names_if_possible(string, pattern, out)
 }
 
 like_to_regex <- function(pattern) {
-  converted <- stri_replace_all_regex(pattern, "(?<!\\\\|\\[)%(?!\\])", "\\.\\*")
+  converted <- stri_replace_all_regex(
+    pattern,
+    "(?<!\\\\|\\[)%(?!\\])",
+    "\\.\\*"
+  )
   converted <- stri_replace_all_regex(converted, "(?<!\\\\|\\[)_(?!\\])", "\\.")
   paste0("^", converted, "$")
 }