+++ R/word.R
@@ -50,7 +50,7 @@ word <- function(string, start = 1L, end = start, sep = fixed(" ")) {
 
   # Extract locations
   starts <- mapply(function(word, loc) word[loc, "start"], words, start)
-  ends <-   mapply(function(word, loc) word[loc, "end"], words, end)
+  ends <- mapply(function(word, loc) word[loc, "end"], words, end)
 
-  str_sub(string, starts, ends)
+  copy_names(string, str_sub(string, starts, ends))
 }